move all layout components into layout folder
